feat(monitoring): add support for ingress-nginx
This modification adds a new section to monitoring for alerts, dashboards,
and rules associated with ingress-nginx. This provides additional visibility
into communication coming into the cluster and can help identify when ingress-
level features (such as rate-limiting) kick in.

This change also updates the service catalog to include relavant links for
services who use the ingress-controller for communication.