feat: add taky docker server
taky is an open source, Python implementation of a TAK server. Of
the existing implementations, this has been the only one that has
run out of box. However, taky can be difficult to use as it doesn't
operate on a database. Redis can be used, but it defaults to using
the file system for most services.