feat(charts/12factor): add metrics support
This change supports including a PodMonitor as part of the deployment
configuration. In addition to that, it iterates on how the ingress
port was managed and included some reasonable defaults to reduce
downstream configuration. Finally, this change also included a handful
of examples used to snapshot renderings of the chart.

Resolves https://github.com/mjpitz/mjpitz/issues/82